requirements.txt
apache-beam[gcp]
google-cloud-pubsub